Run `git cl format --full` on safe_conversions_impl.h
It had a large-ish number of non-standard-formatted lines. This caused
a problem in another CL where I try to move this file: `git cl format`
wants to reformat all these lines, which in turn causes the rename
detection to treat it as a delete+add, making reviewing hard.

Better to just reformat it in a separate CL and get it over with.

BUG=none

Change-Id: I619f0454546e8a55fba58da08073da9bb1d06207
Reviewed-on: https://webrtc-review.googlesource.com/20865
Reviewed-by: Danil Chapovalov <danilchap@webrtc.org>
Commit-Queue: Karl Wiberg <kwiberg@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#20582}