Make hbos@webrtc.org OWNER of peerconnection*.
hbos knows and makes changes to the webrtc-pc spec[1] and works on
making Chromium's RTCPeerConnection spec-compliant. This includes
knowing and interacting with WebRTC-layer PeerConnection/Interface and
sometimes making changes to it.

hbos would like to share the peerconnection* ownership responsibilty as
it is relevant and owning it will speed up some of the process.

[1] https://w3c.github.io/webrtc-pc/

Bug: None
NOTRY: True
Change-Id: I8f419b7fc6c7fcf19951aa3f304769c915300d1b
Reviewed-on: https://webrtc-review.googlesource.com/21327
Reviewed-by: Tommi <tommi@webrtc.org>
Reviewed-by: Peter Thatcher <pthatcher@webrtc.org>
Commit-Queue: Henrik Boström <hbos@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#20649}