Remove webrtc_tests.
This target is just confusing, since it runs essentially random tests
out of the test binaries it's depending on. This CL makes the top
GN file structured like Chromium's is.

A possible secondary function is to act as a link test (maybe to catch
twice-defined symbols etc), but that seems highly speculative.

Bug: None
Change-Id: Ic8a36ffe96408312e602a58f8806b0694fedaad6
Reviewed-on: https://webrtc-review.googlesource.com/27380
Commit-Queue: Patrik Höglund <phoglund@webrtc.org>
Reviewed-by: Oleh Prypin <oprypin@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#21028}