Bug Fix: WebRTC Unity Plugin Audio One Way
When audio_only is on for the webrtc unity plugin, there is a bug that
the audio from hologram cannot be heard at the remote side.

Actually we found the audio is transmitted to the remote side, but the
remote side wants video data also to playout everything. So without
video data, the remote side will drop all the audio data.

Thus, on the hologram (using webrtc unity plugin) side, we should not
hook up a dummy camera, but instead we should use media constraint to
request the remote side to send video data.

This CL fixes the bug.

Bug: webrtc:8555
Change-Id: I21ddda65185b645088aa4ac15f47b3f8ffad1873
Reviewed-on: https://webrtc-review.googlesource.com/24680
Commit-Queue: Qiang Chen <qiangchen@chromium.org>
Reviewed-by: George Zhou <gyzhou@chromium.org>
Cr-Commit-Position: refs/heads/master@{#21094}