Generate iOS framework umbrella header.
Instead of keeping the umbrella header in sync manually and needing
ifdefs to make it include the correct headers depending on platform,
generate it based on the headers we include in the framework target.

Can also be used to only include internal software codec headers when
compiling with support for them.

Bug: webrtc:7925
Change-Id: I63f97af1efc8710cfd62d527fcb343fed05daae2
Reviewed-on: https://webrtc-review.googlesource.com/38702
Commit-Queue: Anders Carlsson <andersc@webrtc.org>
Reviewed-by: Patrik Höglund <phoglund@webrtc.org>
Reviewed-by: Kári Helgason <kthelgason@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#21613}