Allows the application to set active simulcast streams.
Currently all simulcast streams are set as active by default. This
update takes the values of the rtp encoding parameters and wires those
values down to the VideoSendStream and VideoStreamEncoder, so that the
appropriate simulcast streams can be turned off and on. This includes
adding more application specific controls in the EncoderStreamFactory.

Bug: webrtc:8653
Change-Id: Iaa7da3209cea0f0db72543981a319e319705cb00
Reviewed-on: https://webrtc-review.googlesource.com/47245
Reviewed-by: Erik Språng <sprang@webrtc.org>
Reviewed-by: Peter Thatcher <pthatcher@webrtc.org>
Commit-Queue: Seth Hampson <shampson@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#21920}