Added some margin to ramp down target in perf test.
The bit rate target for ramp down in was set equal to the simulated
capacity. Expected behavior of an estimator is to achieve an estimate
near the true value but not always the exact value. Adding a margin
allows from noise in the measurement while still testing for the desired
behavor.

Bug: webrtc:8878
Change-Id: I18fb6c9704bf08e58ee08ce6c85abee2eaa08356
Reviewed-on: https://webrtc-review.googlesource.com/57080
Reviewed-by: Niels Moller <nisse@webrtc.org>
Reviewed-by: Stefan Holmer <stefan@webrtc.org>
Commit-Queue: Sebastian Jansson <srte@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#22171}