Change place of UMA logging in AudioMixer.
And fix typo in UMA metric.

We have this pattern in the FrameCombiner component of the AudioMixer:

  if (number_of_streams <= 1) {
    // Copy or fill with zeros.
    return;
  }
  // Mix and limit
  LogMixingStats(/* args */);

When there is only one remote stream, info about active streams and
sample rate is not logged. This CL moves the call to log stats before
the 'return'.

Bug: webrtc:8925
Change-Id: I7b54f61f628273631909dafbfafa21e155e18d4a
Reviewed-on: https://webrtc-review.googlesource.com/62860
Reviewed-by: Alessio Bazzica <alessiob@webrtc.org>
Commit-Queue: Alex Loiko <aleloi@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#22493}