Reduce implementations of CallStatsObserver by 2.
Both of these implementations are added as observers by
VideoReceiveStream, which is also an observer.

Instead, VideoReceiveStream now forward the necessary information
when the stats are updated as is already done for |frame_buffer_|.

Fewer observers registered, simplifies the code a little bit,
reduces synchronization when registering/unregistering observers
and reduces the set of existing implementations.

(I'm changing call_stats.* a little bit, so this is a by-the-way change)

Bug: none
Change-Id: I2d40b80a60e5ca632293c5118ac3b1a35d48ea40
Reviewed-on: https://webrtc-review.googlesource.com/64524
Reviewed-by: Magnus Flodman <mflodman@webrtc.org>
Commit-Queue: Tommi <tommi@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#22594}