roll_deps: Accept any prefix (like 'git_revision:'), not only 'version:' for CIPD
`gclient setdep` was changed in https://chromium-review.googlesource.com/1123940
to support any prefix as well, but note that that was a backwards incompatible
change, because it now requires the prefix to be passed. So we just stop stripping
the prefix in this CL.

Also clarify the error when a CIPD dep is present in WebRTC and missing in Chromium.

No-Try: True
TBR: phoglund@webrtc.org
Bug: webrtc:9470, chromium:858978
Change-Id: I5e42bbda04db37a628a0ac1de69667b9a30dd793
Reviewed-on: https://webrtc-review.googlesource.com/86280
Commit-Queue: Oleh Prypin <oprypin@webrtc.org>
Reviewed-by: Mirko Bonadei <mbonadei@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#23814}