Removes pause check from RoundRobinPacketQueue.
This CL removes a check in RoundRobinPacketQueue::FinalizePop. This
check will trigger if a the pause state is changed in PacedSender while
a packet is sent. This is a rare occurrence but would yield flaky
behavior. The check should not be required for the code to function
since the paused state is not read in FinalizePop other than for this
check.

Bug: webrtc:9586
Change-Id: Ib9476168eb637dc2f9710d0592bed92c4b03dacb
Reviewed-on: https://webrtc-review.googlesource.com/92090
Reviewed-by: Philip Eliasson <philipel@webrtc.org>
Commit-Queue: Sebastian Jansson <srte@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#24197}