Allow send bitrate < start bitrate in RampUpTest.
Primarily, this is intended to reduce flakyness of
RampUpTest.AudioTransportSequenceNumber. We shouldn't expect audio
send rate >= 300 kbps at all time in these tests. And in general, if
it's at all relevant to test that bitrate doesn't drop below the start
bitrate, a perf test isn't the right place for that.

A run of

./third_party/gtest-parallel/gtest-parallel  -r 1000 -w 1000 \
   --gtest_filter=RampUpTest.AudioTransportSequenceNumber \
   out/Release/webrtc_perf_tests

passes when I ran it locally after this change, but fails around 4 out
of 1000 times before the change.

Bug: webrtc:8878
Change-Id: I08614ce5683c9ba6fe4b72bfde83e6a81445a59b
Reviewed-on: https://webrtc-review.googlesource.com/96900
Reviewed-by: Sebastian Jansson <srte@webrtc.org>
Commit-Queue: Niels Moller <nisse@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#24523}