Introduce QualityAnalyzingVideoEncoder and QualityAnalyzingVideoDecoder.
This encoder will be used to inject VideoQualityAnalyzerInterface into
VideoEncoder, so it will be able to measure its metrics and also trace
frames from capturing on one peer side to rendering on another peer side.
The decoder will be used for the same purpose but in VideoDecoder pert.

Bug: webrtc:10138
Change-Id: Idf719753e3c0b3b1369ff206365bf0558705eb98
Reviewed-on: https://webrtc-review.googlesource.com/c/117363
Commit-Queue: Artem Titov <titovartem@webrtc.org>
Reviewed-by: Peter Slatala <psla@webrtc.org>
Reviewed-by: Ilya Nikolaevskiy <ilnik@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#26381}