Add thread safety annotations for some more PeerConnection members (part 5)
Plus all the annotations that were necessary to make things compile
again.

We needed a special twist for call_. The value it points to is owned
by the worker thread, but the signal thread needs to read the pointer.
We could have made the pointer const, except that we explicitly reset
it in the destructor (in an invoke to the worker thread).

Bug: webrtc:9987
Change-Id: I31f024547f4be0e50967133b0d452c80ae38d7ed
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/128863
Commit-Queue: Karl Wiberg <kwiberg@webrtc.org>
Reviewed-by: Steve Anton <steveanton@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#27278}