Fixing issue with creating StreamParams when track id is not signaled.
Current logic requires a stream id and track ids when creating a stream
that does not have SSRCs signaled.
This change removes the requirement for stream ids. The requirement for
track id is softer, as one should be generated when it is not present.

Bug: webrtc:10551
Change-Id: Ibc0cc181c6b18efa8394b6c0e4820e3a13da70c4
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/133080
Reviewed-by: Steve Anton <steveanton@webrtc.org>
Commit-Queue: Amit Hilbuch <amithi@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#27633}