Adding VoIP specific channel adjustments
This CL introduces a channel remapping for the mono input case that
is more tailored to the VoIP usecase.

The CL contains a kill-switch that can be used to fall back to the old
mapping behavior in case a need for this is perceived.

Bug: chromium:1027117
Change-Id: Idaaba6eac952e6436beaaf5a1a697cfab8f63286
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/160280
Commit-Queue: Per Åhgren <peah@webrtc.org>
Reviewed-by: Gustaf Ullberg <gustaf@webrtc.org>
Reviewed-by: Sam Zackrisson <saza@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#29877}