[Stats] Include fecPackets[Reeceived/Discarded] in Members()
This refers to modern getStats() only. The metrics has been implemented
for a while in C++ but was accidentally not included in the Members()
list, meaning they were not exposed in lists (including exposure in
Chrome/JavaScript).

The Chromium whitelist already include them.

TBR=hta@webrtc.org

Bug: webrtc:11317
Change-Id: I0c3ee9c552975fc37db2d87196c66e662c994aed
Reviewed-on: https://webrtc-review.googlesource.com/c/src/+/167530
Reviewed-by: Henrik Boström <hbos@webrtc.org>
Commit-Queue: Henrik Boström <hbos@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#30391}